Use gint64 instead of long to be sure of the size
The size of type `long` can vary on different platforms. In addition,
the size of type `time_t` can vary as well. So be sure to cast `tv_sec`
to a known size.

On most platforms, `sizeof(time_t) == sizeof(gint64)`, so we should be
fine, but let's be safe.